+++ boofuzz/connections/iserial_like.py
@@ -4,17 +4,22 @@ from future.utils import with_metaclass
 
 
 # abc.ABCMeta is the metaclass in both python 2 and 3
-class ITargetConnection(with_metaclass(abc.ABCMeta, object)):
+class ISerialLike(with_metaclass(abc.ABCMeta, object)):
     """
-    Interface for connections to fuzzing targets.
-    Target connections may be opened and closed multiple times. You must open before using send/recv and close
-    afterwards.
+    A serial-like interface, based on the pySerial module,
+    the notable difference being that open() must always be called after the object is first created.
+
+    Facilitates dependency injection in modules that use pySerial.
+
+    .. versionchanged:: 0.2.0
+        ISerialLike has been moved into the connections subpackage.
+        The full path is now boofuzz.connections.iserial_like.ISerialLike
     """
 
     @abc.abstractmethod
     def close(self):
         """
-        Close connection.
+        Close connection to the target.
 
         :return: None
         """
@@ -32,34 +37,22 @@ class ITargetConnection(with_metaclass(abc.ABCMeta, object)):
     @abc.abstractmethod
     def recv(self, max_bytes):
         """
-        Receive up to max_bytes data.
+        Receive up to max_bytes data from the target.
 
         :param max_bytes: Maximum number of bytes to receive.
         :type max_bytes: int
 
-        :return: Received data. bytes('') if no data is received.
+        :return: Received data.
         """
         raise NotImplementedError
 
     @abc.abstractmethod
     def send(self, data):
         """
-        Send data to the target.
+        Send data to the target. Only valid after calling open!
 
         :param data: Data to send.
 
-        :rtype int
         :return: Number of bytes actually sent.
-        """
-        raise NotImplementedError
-
-    @abc.abstractproperty
-    def info(self):
-        """Return description of connection info.
-
-        E.g., "127.0.0.1:2121"
-
-        Returns:
-            str: Connection info descrption
         """
         raise NotImplementedError