+++ boofuzz/connections/udp_socket_connection.py
@@ -8,11 +8,14 @@ import sys
 
 from future.utils import raise_
 
-from . import base_socket_connection, exception, ip_constants
+from boofuzz import exception
+from boofuzz.connections import base_socket_connection, ip_constants
 
 
 class UDPSocketConnection(base_socket_connection.BaseSocketConnection):
     """BaseSocketConnection implementation for use with UDP Sockets.
+
+    .. versionadded:: 0.2.0
 
     Args:
         host (str): Hostname or IP adress of target system.